Avoid input cache when resized (#28)
C++ test copied from #27 

If the producer of a resized tensor is an input cache and we need to
promote it to global memory, just do not cache the input but directly
read from the global memory input. It doesn't make any sense to cache a
global memory input in global memory.

This fixes the issue of #27, which is due to the grid sync inserted for
an input to pad, where the input is a copy of a fusion input.