Optionally traverse Expr attributes (#50)
Add `traverse_attributes` optional flag to `IterVisitor` routines to
traverse into attributes as well. The original motivation is that
`ValReplacementMutator`, used by `replaceSymbolicSizes`, needs to apply
replacements to the left and right expand parameters of `Resize`. This
hasn't been an issue but is exhibited as an error with a small
optimization (#51).

One caveat is that the traversal doesn't visit attributes of type
`Attribute` as it's a template type and the dispatch system doesn't
support dispatching `Val`'s to handlers for template types. Not a
problem for the Resize traversal, but may need to be revisited.