Simplify IterDomain extents if possible (#51)
Simplify extents of `Resize` output IterDomains if possible. Each resize
output domain has an extent of `in->extent() + left_expansion +
right_expansion`. In the case of `slice`, with `start` and `stop`
offsets, the expansion parameters are `-start_offset` and `stop_offset -
in->extent()`, respectively, so the overall extent is `in->extent() -
start_offset + (end_offset - in->extent())`. Of course, this can be
simplified to `end_offset - start_offset`.

Unexpectedly, this simplification exposed a problem with
`ValReplacementMutator`, which is used by `replaceSymbolicSizes`. The
underlying root cause is that IterVisitor does not traverse into
attributes, such as the expansion parameters of `Resize`. It's fixed in
#50.

This also fixes an error encountered with a recently added Python test,
`test_nanogpt_split_mha_linears`. A C++ version of the test is added.

---------

Co-authored-by: Gao, Xiang <qasdfgtyuiop@gmail.com>