Allow chains of reshape (#60)
Previously, fusions like
[this](https://github.com/NVIDIA/Fuser/pull/60/files#diff-a8f5333aa3f2d21440b3cea429bb2a588ed583f4d05486063ef1dc1a30996df9R2411)
are segmented due to a limitation of `DomainMap`.

It seems there's no impact to the existing tests and benchmarks. No
failure with the tests and benchmarks. Dumped all CUDA generated kernels
from the benchmarks and compared before and after this PR. Nothing
changed.

This is part of the fix for #58

---------

Co-authored-by: Gao, Xiang <qasdfgtyuiop@gmail.com>