Ignore Resize ops when validating all ID uses are exactly mapped. (#64)
Resize ops are not replayed, so they don't need to be exactly mapped

Previously, `FusionSliceForNanoGPT3_CUDA` was segmented as the `resize`
ops are not exactly mapped since they have the different expansion
arguments. Since those `resize` ops are part of rfactor transformations,
they were detected as conflicting rfactor transformations. However,
unlike the `split` and `merge` used by `reshape`, `resize` ops are not
replayed, so they don't need to be uniform.

This is also part of the fix for #58. Looks like the Python example is
not segmented anymore, although I suspect there's still something need
to do for `permute`.