CTA Swizzles (#87)
Adds a CTA swizzle to change the order in which the tiles of the output
matrix are processed.
This swizzle increases data reuse from A and B, when iterating over
gridDim.x. Turns out that CTAs are launched in practice by iterating
over gridDim.x first (order is unspecified though, it just happens to
behave the same). As a result, the current wave will contain CTAs that
compute square sub-matrices of C, and so, increase L2 hit rate.

Best factor seems to be 4. This will be part of the heuristics. Setting
the factor to 1 disables this swizzle.

On a 8192x8192x8192 matmul with default config, the speedup is about
20%.

An extreme example is following case: `MNK = 6144 6144 6144, layout=NT
stages=0, cta_tile = 32 32 128, warp_tile = 16 16 128, instruction_tile
= 16 16 16` where the runtime drops from 12.4 ms to 7.28ms !

Thank you @zasdfgbnm for the help. 
Values measured on NVIDIA A100 SXM4 80 GB

---------

Co-authored-by: Gao, Xiang <qasdfgtyuiop@gmail.com>