Add nextafter op (#44)
The `nextafter(x, y)` operation provides the nearest distinct
representable floating point value to `x` between `x` and `y`. In CUDA
these are obtained with the builtins `nextafter` and `nextafterf`. Other
types such as bfloat16 are not directly supported, though PyTorch has
implemented that case based on some code in musl:
https://github.com/pytorch/pytorch/pull/61829/files#diff-ece04c31934b3504382e10ed3e9a69f03ffabd81ad1a2a890aab19b1642f53c0R120.

In PyTorch, the torch.nextafter function is defined for any pair of
arguments which would normally be promoted to either float32 or float64.
So arguments which are both ints, bools, complex, or half-precision
floats are not supported. This PR implements a binary op macro with a
TypePromotionConfig that enforces that rule.

This is a translation/update of
https://github.com/csarofeen/pytorch/pull/2510.

---------

Co-authored-by: Jacob Hinkle <jhinkle@nvidia.com>