Add defer_inserting_tv_uses flag to addUse
This fixes a bug where exprs were never being added to TensorView
uses(), since addUse() was called during resetTvUses(), but addUse only
invalidated the uses and did not add them. We can now force the adding
(which should be done during updating) by setting this flag to false
instead of the default.