Workaround for no-input fusion with reduction (#143)
The reduction and persistent schedulers assume a fusion has inputs, and
there's no check in their `canScheduleCompileTime`, so they just fail
when scheduling a fusion if there's no input. This PR avoids the failure
by rejecting the schedulers if no input is available.

Here's an example fusion, which is a segment of the fusion defined by
the repro #132.

```
T6_l[ iS12{4}, iS13{4} ]
   = full({4, 4}, ( (double)(0) ));
T7_g[ iS14{4}, rS15{4} ]
   = reduction( T6_l[ iS12{4}, iS13{4} ], op = add, initial value = double(0), allreduce = false )
```

This is a WAR. Ideally, we should be able to support reductions with no
input. That's not addressed in this PR.

Closes #132