Cleanup FusionExecutor::runFusion (#139)
Cleans up `FusionExecutor::runFusion`. The main motivation is to remove
the repeated code of buffer allocations. Previously the execution flow
looks like:

```
if (cached-executor-entry-available) {
  allocate-outputs-using-the-executor-entry
  allocate-other-buffers-using-the-executor-entry
} else {
  recompile-the-kernel-if-necessary
  allocate-outputs
  allocate-other-buffers
  save-allocation-info-in-executor-entry
}
```

With this PR:

```
if (!cached-executor-entry-available) {
  initializing-executor-entry
  save-executor-entry-for-future-use
}

allocate-outputs-using-the-executor-entry
allocate-other-buffers-using-the-executor-entry
```

I haven't touched functions that are used by
`FusionKernelRuntime::dryRunKernelWithInput`, which I believe is used
for async compilation, and I don't know how it's implemented, also
there's another related PR on async compilation by Ryan (#130)

This is also in part for Michel's #108