Refactor transpose and set (#148)
This cleanup comes to my mind during @jacobhinkle's talk (thank you so
much for having this talk) where he mentioned that transpose does not
use rfactor domain. I think it makes more sense to model transpose as a
set with rfactor domain in its consumer, and I think this new way would
help me model `ldmatrix.T` and matmul's `NN` support better.

This refactor does the following:
- Remove `UnaryOpType::Set`. Set is now a `LoadStoreOp` with
`LoadStoreOpType::Automatic`. This get rid of a lot of special handling
for set.
- If the consumer of a `LoadStoreOp` contains an rfactor domain, then
this load store op has fused permutation. (I think view can be modeled
the same way, but this is not the purpose of this PR).
- `TransposeOp` is removed completely. Permutation/transpose is now a
`LoadStoreOp` with rfactor in consumer.
- There is no more special handling for transpose in
`PairwiseRootDomainMap` and `ComputeAtRootDomainMap`
- In the future, I will consider modeling `ldmatrix.T` as an operation
with fused transpose, I have not carefully investigated this idea yet,
but I believe this would make it more straightforward to support `NN` in
matmul.