Remove TypeHash (#154)
This is a minor PR removing `TypeHash` from the codebase. This was a
workaround (introduced in
https://github.com/pytorch/pytorch/pull/35917/commits/c5f805590d69a83729a524be5c39db319209e7e5)
for the lack of support for using `enum class`es like `ValType` as keys
in STL hash-tables like `std::unordered_map`. See
https://stackoverflow.com/questions/18837857/cant-use-enum-class-as-unordered-map-key
for more discussion on the general issue. As noted in that issue, this
was resolved in C++14 (see
https://www.open-std.org/jtc1/sc22/wg21/docs/lwg-defects.html#2148) so I
think it should no longer be necessary as NVFuser has even moved on to
C++17 by now.

Co-authored-by: Jacob Hinkle <jhinkle@nvidia.com>