Add `reduction_factor` and `reorder` scheduling ops to Python API (#162)
This PR does 3 things:
1. Adds `reduction_factor` scheduling op. I decided to name explicitly
spell out reduction instead say `rFactor` in the hopes of making its
purpose more obvious.
2. Adds `reorder` scheduling op.
3. Augments the `define_tensor` operation of the `FusionDefinition` to
allow for the tensor dimensions to be defined as concrete sizes in
addition to `-1` (symbolic) and `broadcast` in order to test
`reduction_factor` without having an API for parallelization. Without
parallelization `reduction_factor` wants to create an intermediate
tensor where the size of the local buffer is unknown.