Remove non-dunder op methods from python tensors/scalars (#134)
The methods such as `tv.sum()` is marginally more convenient than the
equivalent `fd.ops.sum(tv)`, but requires a lot of boilerplate in
`python_bindings.cpp`. In order to reduce the maintenance burden, this
PR removes all of these types of methods. We do keep the "dunder"
methods such as `__add__` so that we can perform `tv1 + tv2` and similar
operations.