Minor cleanup of HaloInfo (#202)
@jacobhinkle found `HaloInfo::buildConcreteHaloExtentMap` is a major
overhead in fusion lowering. This PR should fix it.

`HaloInfo` previously kept a permissive map, which shouldn't be
necessary as it's only used in lowering. Also, analysis for LoopIndexing
can be skipped when no root ID is extended with halo.

Note that overall rewrite of halo handling is needed, but it's not high
priority at this moment.