Avoid DCE in fuseWarpReduce when no rewrites happened (#192)
The `fuseWarpReduce` call in lowering rewrites single-warp reductions
with broadcast consumers as allreduces. After that, it runs a custom DCE
pass that aims to remove any replaced broadcasts and associated
allocations. This DCE pass was actually removing an allocation in the
example from #151 where the only use for the allocation is in an index
expression, causing a compilation error.

The current approach is just to short-circuit this function if no
broadcasts are removed. However, as noted in the previously-existing
comments, rewriting and DCE passes like this are better done prior to
this point, on the Fusion graph.

Also note that in debugging this, I realized it is quite hard to measure
the effect of `fuseWarpReduce`. I tried commenting the function and I
think all our tests still ran. It may be the case that we do not
explicitly test rewrites of single-warp reductions, or that single-warp
reductions are already being covered by the previous Fusion-based
rewriting pass called `fuseReductionsAndBroadcasts` which covers grid
reductions. A more detailed look at refactoring these two passes and
potentially removing `fuseWarpReduce` will be left to another PR.

Fixes #151

---------

Co-authored-by: Naoya Maruyama <naoyam@users.noreply.github.com>