Fixing up managed data.
Currently there is a problem where the managed data is continuing to be
managed after it is no longer valid (i.e. when we concretize the
fusion). That leads to a segfault as the symbolic reshaped tensorviews
are freed during concretization. The fix is to add a new facility to
Fusion which lets us "unmanage" data. I'll include that in a separate
commit on this PR, since we may want to save it for a separate PR.