Better validation of domains when creating TensorDomain (#231)
This is for #225 

A `TensorDomain` is created with a list of IterDomains representing a
root domain, and optionally another list of IterDomains for an rfactor
domain and also one more list for the leaf domain.

Currently, we do some error check using `IterVisitor::getInputsTo`, but
that's not always sufficient. We need to make sure that the root domain
and the other domains represent the same logical space without any
overlap.

For example:

```
IterDomain* id0 = IterDomainBuilder()....build();
IterDomain* id1, id2 = IterDomain::split(id0, 4);
```

Here, it'd be valid to have a root and leaf domain as follows:

```
Root: id0
Leaf: id1, id2
```

Suppose `id2` is further split as:

```
IterDomain* id3, id4 = IterDomain::split(id2, 5);
```
Then, this is invalid:

```
Root: id0
Leaf: id1, id2, id3, id4
```

`id3` and `id4` are disjoint, but they overlap with `id2`. To represent
the original space using the new leaf IDs, `id2` shouldn't be included,
however, the current error check won't detect these kinds of invalid
domains.

This PR replaces the existing error check with more complete analysis.
As long as all IDs are non-symbolic, it should be able to determine
whether a list of initial IDs and another list of derived IDs represent
the same space. It also checks if the derived IDs are disjoint.

Since the error check is always used when TensorDomain is created, I
consider it well tested, so I just added one unit test for testing some
corner cases.