Refactor definition of Turing and Ampere matmul and add NN support (#234)
This is the most important but also the easiest step for matmul NN
support. It changes the fusion definition of Turing & Ampere matmul into
the "Easiness of lowering" format as described in
https://github.com/NVIDIA/Fuser/issues/203. This refactor allows me to
add NN support for Turing & Ampere.

Unfortunately, the user interface is made worse by this PR: originally,
users only need to specify the input format (TT, TN, NT) when defining
matmul. But now, users also need to specify the architecture when
defining matmul. This is undoubtedly unpleasant, but I believe we must
accept this regression in order to be able to do incremental
development.