Support dynamic reshapes in FusionExecutorCache (#215)
# Background

#24 introduces support for dynamic shape arguments for the `reshape`
command. When the shape cannot be fully inferred at runtime (defined
here as having non-constant scalars present), then output `IterDomain`s
are marked as having `IterType::Symbolic`. These are not allowed during
lowering, and one must "concretize" the `Fusion` by binding some
concrete integers to make the associated extents constant, before
scheduling/lowering/execution.

Given a compute definition for a `Fusion`, the `FusionExecutorCache`
maps from inputs to `FusionKernelRuntime`s. This means that when passed
inputs with new sizes or on a new device, `FusionExecutorCache` will
determine whether it is safe to use a previously-determined index type
(int32 or int64) and scheduler heuristics, and reuse
`FusionKernelRuntime` if possible, avoiding
resegmentation/rescheduling/recompilation. With the introduction of
dynamic reshapes, we must ensure that reuse occurs only when the inputs
are compatible with the cached concretized transforms.

# Approach

The current PR triggers a resegmentation and full rescheduling of the
entire original `Fusion` whenever the inputs lead to a new set of
concretized transforms. For a small `Fusion` this is reasonable.
However, for large segmented `Fusion`s, this will be wasteful, as
potentially only a single segment may require rescheduling/recompilation
due to changing dynamic transforms. Re-using all concrete segmentation
groups while rescheduling dynamic ones in such a case would be
preferable, but we must be careful since (I think) this has the
potential to require a resegmentation. That is, changing the transforms
in a single segmentation group might invalidate the segment and require
resegmentation of that segment and that may in turn result in a less
than optimal global segmentation. Instead, we currently just resegment
starting from scratch.

---------

Co-authored-by: Naoya Maruyama <nmaruyama@nvidia.com>