Add take_along_axis to Python frontend (#254)
This exposes the `take_along_axis` backend op introduced in #206 to the
frontend. It behaves very similarly to `gather` including using the same
order of arguments (unlike the corresponding backend ops). When calls to
both are valid, the returned tensors will match, but `next_after_op`
requires that all non-index dimensions in the index tensor argument
match their counterparts in the value tensor argument.