Add internal-only Python API for experimental matmul operations based on layout (#272)
The following operations are added to the Python API:

-  `ops._matmul_nn`
-  `ops._matmul_nt`
-  `ops._matmul_tn`
-  `ops._matmul_tt`

These operations have the following restrictions:

- They are dependent on sizes M, N, and K sizes defined in multiples of
8. If they are not defined by multiples of 8, currently, codegen returns
a tensor of zeroes.
- FP16 inputs are required but the output tensor cannot be casted back
to FP16. When attempting a cast, an error occurs in segmentation.
- The input tensors defined in Pytorch need to be contiguous in the
intended layout. Using the inputs as defined for Pytorch does not work.

For example, in order to define an NN layout example you need to do the
following:
`M = 24, N = 16, K = 8`
For nvFuser:
```
nvf_inputs_nn = [
    torch.randn(8, 24, device="cuda", dtype=torch.float16),
    torch.randn(16, 8, device="cuda", dtype=torch.float16),
]
```
Eager. Pytorch expects `input1 = [M, K] and input2 = [K, N]` which is
why the transposes are necessary.
```
eager_inputs_nn = [
    torch.randn(8, 24, device="cuda", dtype=torch.float16).transpose(0, 1),
    torch.randn(16, 8, device="cuda", dtype=torch.float16).transpose(0, 1),
]
```

---------

Co-authored-by: kevin.stephano@gmail.com <me@dgx1v-loki-25.nvidia.com>