take_along_axis with non fusion inputs (#250)
This PR extends the scheduler support of `take_along_axis` so hat its
input tensor no longer needs to be a fusion input. Much like `pad` and
`slice`, when there's a mismatched parallelization between the input and
output tensors of a `take_along_axis` op, the input tensor is
automatically placed in a shared memory, which may be
`MemoryType::Shared` or `MemoryType::Global` depending on how the op is
parallelized.

The current scheduling policy is to make `take_along_axis` handled as if
it were a pointwise op, even though it isn't. It may not be optimal, but
I don't think we want to have another scheduler just for these
gather-like ops, but rather these ops should be fused with reductions
and normalizations if possible, so I think it makes most sense to extend
the existing scheduler to be able to fuse `take_along_axis`.

To make `take_along_axis` schedulable like normal pointwise ops, I
extended the transform replay and parallelization propagation so that
propagations are permissively done across index-accessed domains. For
example, when we have:

```
auto tv0 = [I0, I1];
auto tv1 = take_along_axis(tv0, index_tensor, 1);
```

And:

```
tv1->split(1, 4);
```

Then, `replayPasC` with `tv0` as a producer and `tv1` as a consumer
would replay the split of `tv1` on `tv0->axis(1)`.

This permissive propagation would be fine if the input and output
domains have similar sizes, but otherwise, it may be more efficient to
use different transformations. That is not done in this PR, but rather
the focus is to make these ops be fused together as much as possible.

The decision of the validity of fusing these gather-like ops mostly just
depends on whether all IterDomains are mapped with scheduler reference
tensors. For example, in a pointwise case like below:

```
auto tv0 = [I0, I1];
auto tv1 = take_along_axis(tv0, index_tensor, 1);
fusion.addOutput(tv1);
```

`tv1` would be picked as the reference tensor of scheduling, and the
scheduler analyzes the fusion if all input tensors are mapped with the
reference. This mapping only meant the exact mapping, so it would fail
for the domain of `I1`. It is now extended to allow mappings through
index-based accesses.

In the case of the reduction and normalization schedulers, the output
tensor of a reduction expr is used as the reference. When
`take_along_axis` is used after the reduction, we should be able to
schedule it just like any other pointwise ops. When `take_along_axis` is
used before the reduction, if the indexed domain becomes a size-1 domain
and is squeezed, we can no longer propagate anything from the reduction
tensor to the indexed domain, so in that case the reduction scheduler
needs to reject the fusion. It is also the same in the normalization
scheduler.

This can be similarly extended for `select` and `index_select` but I
haven't considered them yet, although I believe very little change would
be needed for them. `torch_gather` is a little different, though. I
think we should decompose it to `slice` for non-indexed domains and
`take_along_axis`.

I added the cross-entropy test that was originally created by
@jjsjann123. It is currently segmented into two kernels, softmax as one
kernel and `take_along_axis` and the final reduction as another kernel.
Ideally, the `take_along_axis` op should be fused to the preceding
softmax, which I'll address by a follow-up PR.