Move NoOp zero-size check to runtime.
Tests still fail. It's difficult to satisfy NoOp by replacement with
full, unless we also concretize zero-size extents in such a way that
size-zero extents are constant. That would mean moving the check back to
compile time for NoOp, which would be simpler. It would also mean all
Fusions with reductions would be dynamic, as originally discussed
between myself and @naoyam. This would mean adding a concretization step
for each new set of input sizes, but that can be made faster as well,
e.g. #244.