`RelwithDebInfo` build mode support (#311)
The `RelwithDebInfo` build mode can be thought of as a mixture of
release and debug build: it enable optimizations like in release build,
it also keep debugging info such as line info. This is my favorite build
mode, and my default build mode for PyTorch. With this build mode, tests
are running as fast as in release build, while keeping the debugability
close to debug build. Strongly recommend to use :)