Promote fusing select-like ops with producers (#332)
For #278, our schedulers currently can't fuse them into a single kernel,
so a segmentation needs to happen. Currently, it's segmented between
softmax and take_along_axis, just because of the ordering of the
segmenter. However, we want the take_along_axis op to be fused together
with the preceding softmax since then the temporary output from the
first segment would be much smaller, reducing gmem access overhead. In
the case of `[8192, 32768]`, the (logical) I/O cost would be `1 /
32768`.

This PR introduces a simple mechanism to allow preferred fusions in the
segmentation steps. Currently, there's only preference of select-like
ops with producers. "Select-like" here also includes index_select and
torch_gather to size-one domains. In those ops, it's guaranteed that the
size of the consumer tensor is no larger than the lookup tensor, so it
makes sense to fuse them with producers.

Currently, it's only tested with the cross-entropy loss case. The
overall segmentation algorithm would need to go through significant
refactoring, so I don't think making this interface super robust is
worth doing at this moment, and it's likely to be redesigned. For now,
this is very important for the cross-entropy performance.