Defer fusion->hasDynamicTransform() in FEC
This fixes a subtle bug where all C++ tests were passing, but for
example `test_cat` in `test_python_frontend.py` failed. The exact same
test would fail in Python that succeeded in C++. The reason is that in
the Python frontend, a `FusionExecutorCache` is initialized _before_
defining the fusion, i.e. with an empty `Fusion`. Doing the check for
dynamic transforms in the constructor then for `FusionExecutorCache`
lead to always finding a static (empty) transform. Instead, this commit
adds the `isDynamic()` method which will cache its results. It should
only be used somewhere inside `runFusionWithInputs` which indicates that
the definition of the `Fusion` is complete.