Concretize IterDomains in-place instead of building
This change replaces uses of IterDomainBuilder which previously required
registerMutation. This is because replacing an IterDomain in a
TensorDomain requires creation of an entire new TensorDomain and
swapping it out in the enclosing TensorView. This change simplifies that
by introducing a protected method IterDomain::setIterType, and making
DynamicTransformConcretizer a friend of IterDomain. Then we can simply
change the iter_type and not require any changes to the TensorDomain or
TensorView. This also means we do not need to modify the resize()
command since we will not call it in order to create the concretized
IterDomain for resize().