Allocation domain support for fusion inputs (#338)
This PR is stacked on https://github.com/NVIDIA/Fuser/pull/337, please
review that PR first.

This PR adds support for allocation domains in tensor inputs. The ATen
tensor provided to the executor is always in the format of rFactor
domain, and we need to analyze the relationship between rFactor domain
and allocation domain to convert the strides in the format of allocation
domain. The shape is still left in rFactor domain, because things like
`T1.size[1]` in generated code comes from replacing the rFactor domain's
extents with named scalar, so these should be in rFactor domain.