Use aligned block sync in block reduction runtime kernel (#351)
This PR changes how block sync is done in some of the runtime functions.
Presently, we always use `__barrier_sync(0)` as runtime functions may be
called with divergent threads, which isn't optimal if threads are
aligned. This PR extends what we already do for the `BlockSync` IR node
and uses `__syncthreads()` when possible.

The primary change is:

```
template <bool aligned = false>
__forceinline__ __device__ void sync() {
  if constexpr (aligned) {
    __syncthreads();
  } else {
    __barrier_sync(0);
  }
}
```

And for now I only extended `blockReduce` to use the `aligned` template
parameter.

Another major change is in codegen.cpp. I added `std::vector<bool>
aligned_scope_exprs_` to represent if a scope is aligned or not. I
thought maybe we might want to add a bool property to `Expr` to indicate
if it's aligned or not, but it depends on the surrounding scopes, so
adding that property to `Expr` doesn't seem like a good idea.

As part of extending `CudaKernelGenerator`, I also refactored it to use
`kir::IrVisitor`.

This PR only changes `blockReduce`. I'll have follow-up PRs to address
other runtime routines.

In our current benchmarks, there are about 300 cases of `blockReduce`,
of which about 200 are aligned. I'll get performance results as well.

Looks like there's no noticeable perf difference. The below graph
compares the bandwidths of all the benchmarks that now use the aligned
sync vs. the same benchmarks with the default barrier sync. There are
132 benchmarks that contain aligned blockReduce.


![image](https://github.com/NVIDIA/Fuser/assets/363169/ee7378f1-21ff-4e41-a952-217f038a91c0)