Check concretizations for each Expr (#380)
When concretizing `Val* old_val` with `Val* new_val`, we should check
that `new_val` is valid wherever it is used. This is best done by a
check that's internal to each `Expr`'s definition so that it is
self-contained, instead of needing to specify those requirements in the
concretization code or other locations where we might replace vals. This
PR introduces a new method, `Expr::checkConcretization(Val* old_val,
Val* new_val)` which will be called for all `uses()` of replaced `Val`s
during concretization.

Currently, this is only used to check that concretized inputs to
`SqueezeOp` have squeezed IDs that are of type `Broadcast`. The
`SqueezeOp` op checks the concretized `IterType` to ensure we do not
silently concretize to a `Fusion` that squeezes an `Iteration` domain,
which would be invalid.

Note that this PR tests this functionality by enabling the `squeeze` opt
to accept `Symbolic` squeezed IDs. So it fixes #379.