Back to defaulting to no input scalars affecting ID
This better matches the original use. Currently the only use for this
call method is in fusion_cache, enabling user schedules in the Python
frontend. In that case, there should be no symbolic fusions since those
can't be scheduled. So in that case it really makes more sense to not
record any input scalars.