Precompute some concretization information (#310)
Fixes #244.

Introduces `DynamicTransformInitialInfo` which is information obtained
before inputs are available that is used to speed up the actual
concretization information. In particular, graph traversals to find
dynamic ops are only done once, to build this object, after which each
new set of inputs that needs to be concretized only needs to replace
those ops (constant time) followed by traversals to update symbolic
IterDomain types and update TV uses.

The `DynamicTransformInitialInfo` object also holds a _minimal_ set of
input scalars that determine concretization. This is used for minimizing
the cache ID associated with `KernelArgumentHolder` so that input
scalars which do not affect concretization do not mistakenly alter the
key.

We also attach an `ExpressionEvaluator` to
`DynamicTransformInitialInfo`, and warm it up so that concretization is
slightly faster.

Incidentally, this PR also adds a new `fusion_ir_concretized` debug dump
option.